ビルドエラー解消

+++ src/main/kotlin/library/Force.kt
@@ -3,6 +3,7 @@ package library
 /**
  * force * time leap
  */
-data class Force(val value: Float) {
+data class Force(val value: Double) {
+
 
 }
